Divide class into package.
