+++ src/main/java/duke/command/Parser.java
@@ -1,3 +1,7 @@
+package duke.command;
+
+import duke.DukeException;
+
 import java.util.Arrays;
 
 public class Parser {
