+++ src/main/java/duke/command/Storage.java
@@ -1,3 +1,5 @@
+package duke.command;
+
 import java.io.File;
 import java.io.FileWriter;
 import java.io.IOException;
