+++ src/main/java/duke/command/Ui.java
@@ -1,4 +1,5 @@
-import java.util.Arrays;
+package duke.command;
+
 import java.util.Scanner;
 
 public class Ui {
