+++ src/main/java/duke/task/Task.java
@@ -1,7 +1,8 @@
+package duke.task;
+
 public abstract class Task {
 
     boolean checked = false;
-
 
     public void setChecked() {
         this.checked = true;
