+++ src/main/java/duke/task/Todo.java
@@ -1,9 +1,13 @@
-public class Todo extends Task{
+package duke.task;
+
+import duke.DukeException;
+
+public class Todo extends Task {
     private final String title;
 
     public Todo(String command) throws DukeException {
         if (command.isEmpty()) {
-            throw new DukeException("Please use this format: Todo <Activity>");
+            throw new DukeException("Please use this format: duke.task.Todo <Activity>");
         }
         this.title = command;
         System.out.println("added: " + this.toString());
